Fix MSVC C4244 and refactor joint-AAD for Codacy complexity
- Add static_cast<double>() on timer.Elapsed<milliseconds>() calls
  in the joint example (MSVC /WX C4244: int64_t->double truncation).
- Extract RegisterTapeParameters and BuildTemplatedCurves from
  AnalyticJacobian to reduce cyclomatic complexity (3 Codacy medium).
- Hoist the liborBasis_ ACT_365F check out of the per-declaration loop
  in JointSpecEligibleForAnalyticJacobian.
- Fix NOTICE(String_ + ...) calls (XStackInfo_ ctor is private).

750/750 tests pass.

+++ dal-cpp/examples/joint_multi_curve_calibration/joint_multi_curve_calibration.cpp
@@ -404,12 +404,12 @@ int main() {
     timer.Reset();
     const JointMultiCurveCalibrationSpec_ jointSpec = BuildJointSpec(today, ccy, market);
     const JointMultiCurveCalibrationResult_ jointResult = CalibrateJointMultiCurve(jointSpec);
-    const double jointMs = timer.Elapsed<milliseconds>();
+    const double jointMs = static_cast<double>(timer.Elapsed<milliseconds>());
 
     timer.Reset();
     const MultiCurveCalibrationSpec_ stagedSpec = BuildStagedSpec(today, ccy, market);
     const MultiCurveCalibrationResult_ stagedResult = CalibrateMultiCurve(stagedSpec);
-    const double stagedMs = timer.Elapsed<milliseconds>();
+    const double stagedMs = static_cast<double>(timer.Elapsed<milliseconds>());
 
     PrintResidualSummary("joint", jointResult.diagnostics_);
     PrintResidualSummary("staged", stagedResult.diagnostics_);
